entity/*: merged storage.hpp and pool.hpp
